feat: lazy load vault list expanded content
